refactor: rename to features/customAccounts and drain the remaining strays
Follow-up to the previous commit, acting on review feedback.

- features/bankAccounts -> features/customAccounts. The bankAccounts pairing was
  spurious: both components only touch @schemas/customAccounts and the
  custom-accounts routes, while hooks/features/bankAccounts is about linked accounts.
- Tags/TagDimensionsGroup -> @blocks. Its four consumers split across
  bankTransactions and generalLedger, so no single feature domain fits.
- SolopreneurOnboardingBanner -> views/SolopreneurOverview, as a view-private
  component, split into its two banner variants plus a state hook holding the
  OnboardingBannerState enum and the pure reducer.
- ExpensesSummaryCard -> features/profitAndLoss.
- HorizontalBarChart -> @blocks. It composes Legend and recharts into a pattern;
  Legend and @ui/Header stay primitives.

Fixes an incomplete rename from the previous commit: Container name= still emitted
Layer__CategorizationRulesView, so the root overflow rule matched nothing on desktop.
