+++ src/components/ExpandedBankTransactionRow/ExpandedBankTransactionRow.tsx
@@ -40,11 +40,11 @@ import { HStack, VStack } from '@ui/Stack/Stack'
 import { Toggle, ToggleSize } from '@ui/Toggle/Toggle'
 import { ErrorText } from '@ui/Typography/ErrorText'
 import { Span } from '@ui/Typography/Text'
+import { TagDimensionsGroup } from '@blocks/TagDimensionsGroup/TagDimensionsGroup'
 import { BankTransactionCategoryComboBox } from '@components/BankTransactionCategoryComboBox/BankTransactionCategoryComboBox'
 import { BankTransactionFormFields } from '@components/BankTransactionFormFields/BankTransactionFormFields'
 import { BankTransactionReceiptsWithProvider } from '@components/BankTransactionReceipts/BankTransactionReceipts'
 import { MatchForm } from '@components/MatchForm/MatchForm'
-import { TagDimensionsGroup } from '@components/Tags/TagDimensionsGroup/TagDimensionsGroup'
 import { TaxCodeComboBox } from '@components/TaxCodeSelect/TaxCodeComboBox'
 import { CustomerVendorSelector } from '@features/customerVendor/CustomerVendorSelector/CustomerVendorSelector'
