--- src/components/SolopreneurOnboardingBanner/SolopreneurOnboardingBanner.tsx
@@ -1,131 +1,0 @@
-import { useCallback, useContext } from 'react'
-import { Info } from 'lucide-react'
-import { useTranslation } from 'react-i18next'
-
-import { type PlaidHostedLinkConfig } from '@schemas/linkedAccounts/plaid'
-import { useGetAccountingConfiguration } from '@api/businesses/[business-id]/accounting-config/get'
-import { useGetTaxProfile } from '@api/businesses/[business-id]/tax-estimates/profile/get'
-import { useSizeClass } from '@hooks/utils/size/useWindowSize'
-import { LinkedAccountsProvider } from '@providers/LinkedAccountsProvider/LinkedAccountsProvider'
-import { useBankAccountsContext } from '@contexts/BankAccountsContext/BankAccountsContext'
-import { useLayerContext } from '@contexts/LayerContext/LayerContext'
-import { LinkedAccountsContext } from '@contexts/LinkedAccountsContext/LinkedAccountsContext'
-import { Banner } from '@ui/Banner/Banner'
-import { Button as LayerButton } from '@ui/Button/Button'
-import { HStack } from '@ui/Stack/Stack'
-import { PlaidHostedLinkErrorBanner } from '@features/linkedAccounts/PlaidHostedLinkErrorBanner/PlaidHostedLinkErrorBanner'
-
-import './solopreneurOnboardingBanner.scss'
-
-enum OnboardingBannerState {
-  Loading = 'Loading',
-  HostedLinkError = 'HostedLinkError',
-  NoBankAccountsLinked = 'NoBankAccountsLinked',
-  NoTaxProfile = 'NoTaxProfile',
-  Onboarded = 'Onboarded',
-}
-
-const getOnboardingBannerState = ({
-  isLoading,
-  isHostedLinkError,
-  hasLinkedAccounts,
-  hasSavedTaxProfile,
-  isTaxEstimatesEnabled,
-}: {
-  isLoading: boolean
-  isHostedLinkError: boolean
-  hasLinkedAccounts: boolean
-  hasSavedTaxProfile: boolean
-  isTaxEstimatesEnabled: boolean
-}) => {
-  if (isHostedLinkError) {
-    return OnboardingBannerState.HostedLinkError
-  }
-
-  if (isLoading) {
-    return OnboardingBannerState.Loading
-  }
-
-  if (!hasLinkedAccounts) {
-    return OnboardingBannerState.NoBankAccountsLinked
-  }
-
-  if (isTaxEstimatesEnabled && !hasSavedTaxProfile) {
-    return OnboardingBannerState.NoTaxProfile
-  }
-
-  return OnboardingBannerState.Onboarded
-}
-
-const NoBankAccountsLinkedBanner = () => {
-  const { addConnection } = useContext(LinkedAccountsContext)
-  const { isMobile } = useSizeClass()
-  const { t } = useTranslation()
-  const handleLinkBankAccounts = useCallback(() => {
-    void addConnection('PLAID')
-  }, [addConnection])
-  const Icon = isMobile ? null : <Info size={16} />
-  const title = t('linkedAccounts:label.link_your_bank_accounts', 'Link your bank accounts')
-  const description = t('linkedAccounts:label.link_your_bank_accounts_description', 'Linking your bank accounts allows us to load your bank transactions and automatically categorize them.')
-  const Button = <LayerButton onPress={handleLinkBankAccounts} variant='outlined'>{title}</LayerButton>
-  return <Banner title={title} description={description} slots={{ Icon, Button }} />
-}
-
-const NoTaxProfileBanner = ({ onSetupTaxProfile }: Pick<SolopreneurOnboardingBannerProps, 'onSetupTaxProfile'>) => {
-  const { isMobile } = useSizeClass()
-  const { t } = useTranslation()
-  const Icon = isMobile ? null : <Info size={16} />
-  const title = t('taxEstimates:label.set_up_your_tax_profile', 'Set up your tax profile')
-  const description = t('taxEstimates:label.set_up_your_tax_profile_description', 'Configuring your tax profile allows us to provide you with tax estimates and avoid any surprises come tax time.')
-  const Button = onSetupTaxProfile ? <LayerButton onPress={onSetupTaxProfile} variant='outlined'>{title}</LayerButton> : null
-  return <Banner title={title} description={description} slots={{ Icon, Button }} />
-}
-
-function SolopreneurOnboardingBannerInternal({ onSetupTaxProfile }: Pick<SolopreneurOnboardingBannerProps, 'onSetupTaxProfile'>) {
-  const state = useSolopreneurOnboardingBannerState()
-
-  if (state === OnboardingBannerState.Loading || state === OnboardingBannerState.Onboarded) {
-    return null
-  }
-
-  return (
-    <HStack className='Layer__SolopreneurLayout__OnboardingBanner'>
-      {state === OnboardingBannerState.HostedLinkError && <PlaidHostedLinkErrorBanner showRetryButton />}
-      {state === OnboardingBannerState.NoBankAccountsLinked && <NoBankAccountsLinkedBanner />}
-      {state === OnboardingBannerState.NoTaxProfile && <NoTaxProfileBanner onSetupTaxProfile={onSetupTaxProfile} />}
-    </HStack>
-  )
-}
-export type SolopreneurOnboardingBannerProps = {
-  onSetupTaxProfile?: () => void
-  plaidHostedLinkConfig?: PlaidHostedLinkConfig
-}
-
-export function SolopreneurOnboardingBanner({ onSetupTaxProfile, plaidHostedLinkConfig }: SolopreneurOnboardingBannerProps) {
-  return (
-    <LinkedAccountsProvider plaidHostedLinkConfig={plaidHostedLinkConfig}>
-      <SolopreneurOnboardingBannerInternal onSetupTaxProfile={onSetupTaxProfile} />
-    </LinkedAccountsProvider>
-  )
-}
-
-const useSolopreneurOnboardingBannerState = () => {
-  const { businessId } = useLayerContext()
-  const { data: accountingConfiguration, isLoading: isAccountingConfigLoading } = useGetAccountingConfiguration({ businessId })
-  const { data: linkedAccounts, loadingStatus: linkedAccountsLoadingStatus } = useBankAccountsContext()
-  const { isHostedLinkError } = useContext(LinkedAccountsContext)
-  const { data: taxProfile, isLoading: isTaxProfileLoading } = useGetTaxProfile()
-
-  const isTaxEstimatesEnabled = !!accountingConfiguration?.enableTaxEstimates
-
-  const isLoading = isAccountingConfigLoading
-    || (isTaxEstimatesEnabled && isTaxProfileLoading)
-    || linkedAccountsLoadingStatus === 'loading'
-    || linkedAccountsLoadingStatus === 'initial'
-
-  const hasLinkedAccounts = Array.isArray(linkedAccounts) && linkedAccounts.length > 0
-
-  const hasSavedTaxProfile = taxProfile?.userHasSavedTaxProfile === true
-
-  return getOnboardingBannerState({ isLoading, isHostedLinkError, hasLinkedAccounts, hasSavedTaxProfile, isTaxEstimatesEnabled })
-}
