+++ src/components/blocks/HorizontalBarChart/horizontalBarChart.scss
