+++ src/components/blocks/TagDimensionsGroup/TagDimensionsGroup.tsx
@@ -1,7 +1,7 @@
 import { type Tag } from '@schemas/tag'
 import { useLayerContext } from '@contexts/LayerContext/LayerContext'
 import { HStack } from '@ui/Stack/Stack'
-import { TagDimensionCombobox } from '@components/Tags/TagDimensionsGroup/TagDimensionCombobox'
+import { TagDimensionCombobox } from '@blocks/TagDimensionsGroup/TagDimensionCombobox'
 
 import './tagDimensionsGroup.scss'
