+++ src/components/features/customAccounts/CustomAccountForm/CustomAccountForm.test.tsx
@@ -8,7 +8,7 @@ import {
   CustomAccountSubtype,
   CustomAccountType,
 } from '@schemas/customAccounts'
-import { CustomAccountForm, type CustomAccountsFormProps } from '@features/bankAccounts/CustomAccountForm/CustomAccountForm'
+import { CustomAccountForm, type CustomAccountsFormProps } from '@features/customAccounts/CustomAccountForm/CustomAccountForm'
 
 import { post as postCustomAccount } from '@msw/api/businesses/[business-id]/custom-accounts/post'
 import { server } from '@msw/node'
