+++ src/components/features/profitAndLoss/ExpensesSummaryCard/expensesSummaryCard.scss
