+++ src/components/features/taxEstimates/TaxEstimatesSummaryCard/TaxEstimatesSummaryCardContent.tsx
@@ -6,12 +6,12 @@ import { type TaxSummarySectionType, TaxSummaryState } from '@schemas/taxEstimat
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { type DetailData, type SeriesData } from '@ui/Chart/seriesTypes'
-import { HorizontalBarChart } from '@ui/HorizontalBarChart/HorizontalBarChart'
 import { LegendLayout } from '@ui/Legend/Legend'
 import { HStack, Stack, VStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
 import { DetailedChart, type DetailedChartProps } from '@blocks/DetailedChart/DetailedChart'
 import { DetailedTableWithData, NO_OP_INTERACTION_PROPS, NO_SORT_PROPS } from '@blocks/DetailedTable/DetailedTable'
+import { HorizontalBarChart } from '@blocks/HorizontalBarChart/HorizontalBarChart'
 import { resolveCategoryColor, TaxEstimatesSummaryCardMode } from '@features/taxEstimates/TaxEstimatesSummaryCard/constants'
 import { TaxEstimatesSummaryCardEmpty } from '@features/taxEstimates/TaxEstimatesSummaryCard/TaxEstimatesSummaryCardEmpty'
 import { TaxEstimatesSummaryCardNegativeOrZero } from '@features/taxEstimates/TaxEstimatesSummaryCard/TaxEstimatesSummaryCardNegativeOrZero'
