+++ src/views/SolopreneurOverview/SolopreneurOnboardingBanner/solopreneurOnboardingBanner.scss
